add trait VirtioQueueNotifiable for VirtioDevice
Moves queue_notify() to `trait VirtioDevice` because it applies to all
transports:
 - PCI notification register location: configurable, see Virtio-v1.1
     chapter 4.1.4.4
 - MMIO: Virtio-v1.1 chapter 4.2.2, QueueNotify (0x50)
 - Channel I/O for S/390 arch: unsure whether relevant

It has an empty default implementation in `trait VirtioQueueNotifiable`
but adds the possibility to define Device-specific queue_notify()
implementations. Those are required when there is no ioeventfd and
writes to queue_notify actually reach the userspace mmio bus.

This is a breaking change as it requires depedent crates (like devices)
to add a line of code in order to use the empty default implementation
and thus auto-implement the new VirtioDevice trait.